Create README.md
Wanna see some really ugly code? You've come to the right place my son. This app uses timezonedb.com to fetch the time in Kimiliba or whatever that place is called I forgot and don't give a fuck. It says it in the request call. Feel free to improve on this or point out why it's going to fail on Saturday.